Enemy prefabs and pool recycling
Added several new prefabs for enemy variants, varying in max health, type of shot, type of movement and speed.

Now the pool recycling factory script also manages the creation of enemies, storing them when not used (instead of destroying them) and then activating them again when needed. The use of each kind of enemy is managed by groups of enemy prefabs, it is intended to separate enemies based on difficulty level