XML Level Load and Sounds
Now levels can be scripted in a xml file, which will be read by an enemy spawner controller and spawn the indicated type of enemy in the indicated position in the indicated time of appearance.

The player's main actions now have a sound audio that is played when performed. Also added background music.